[audio_core] mix range checks and object lifetimes
Improve boundary checks, integer arithmetic precision, and object
ownership across audio mixing, DSP sampling, and effect loading
pipelines. Specifically:

- Mixer / MixStage: Cap destination frame count in Mixer::Mix during
  gain ramping to *dest.frame_offset_ptr + kScaleArrLen (960 frames),
  ensuring that sampling operations remain strictly within the
  allocated scale_arr buffer during large mix requests (b/521640280).
- MixStage: Widen block size calculation in MixStage::MixStage from
  32-bit integer multiplication to size_t (b/521640620), preventing
  arithmetic overflow when sizing multi-channel output buffers.
- Sampler: Adjust signed modulo remainder and quotient division in
  Sampler::State::AdvancePositionsBy when dest_frames is negative
  (b/525551854), ensuring source_pos_modulo_ always remains positive and
  strictly less than step_size_denominator_.
- EffectV1: Store instance_name_ as an owning std::string instead of a
  non-owning std::string_view (b/521640738), guaranteeing string
  validity across pipeline reconfigurations and extended lifetimes.

New and updated unit test cases:
- MixStageTest.RampingMixExceedsScaleArrLen
- DynamicRange.RampingMixExceedsScaleArrLen
- MixStageTest.BlockSizeMultiplication
- SamplerStatePositionTest.AdvancePositionsNegativeDestFrames
- EffectV1Test.InstanceNameLifetime

Fixed: 521640280
Fixed: 521640620
Fixed: 525551854
Fixed: 521640738
Test: fx test audio_core_unittests
Test: fx test audio_lib_processing_unittests
Test: fx test audio_effects_loader_unittests
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1685895
Commit-Queue: Martin Puryear <mpuryear@google.com>

+++ src/media/audio/audio_core/mix_stage.cc
@@ -71,7 +71,7 @@ MixStage::MixStage(const Format& output_format, uint32_t block_size,
                    std::optional<float> max_gain_db)
     : ReadableStream("MixStage", output_format),
       output_buffer_frames_(block_size),
-      output_buffer_(block_size * output_format.channels()),
+      output_buffer_(static_cast<size_t>(block_size) * output_format.channels()),
       output_ref_clock_(std::move(ref_clock)),
       output_ref_clock_to_fractional_frame_(std::move(ref_time_to_frac_presentation_frame)),
       gain_limits_{
